additions for flux
