+++ src/airunner/data/bootstrap/controlnet_bootstrap_data.py
@@ -119,3 +119,4 @@ controlnet_bootstrap_data = [
 
 if not AIRUNNER_ART_ENABLED:
     controlnet_bootstrap_data = []
+
