+++ src/airunner/data/bootstrap/flux_file_bootstrap_data.py
@@ -1,0 +1,17 @@
+FLUX_FILE_BOOTSTRAP_DATA = {
+    "Flux S": [
+        "model_index.json",
+        "scheduler/scheduler_config.json",
+        "text_encoder/config.json",
+        "text_encoder_2/config.json",
+        "text_encoder_2/model.safetensors.index.json",
+        "tokenizer/merges.txt",
+        "tokenizer/special_tokens_map.json",
+        "tokenizer_config.json",
+        "tokenizer/vocab.json",
+        "tokenizer_2/special_tokens_map.json",
+        "tokenizer_2/tokenizer_config.json",
+        "transformer/config.json",
+        "vae/config.json",
+    ],
+}
