+++ src/airunner/data/bootstrap/pipeline_bootstrap_data.py
@@ -1,126 +1,132 @@
 from airunner.settings import AIRUNNER_ART_ENABLED
 
 
-
 art_pipline_data = [
     {
         "pipeline_action": "safety_checker",
         "version": "SD 1.5",
         "category": "stablediffusion",
         "classname": "diffusers.pipelines.stable_diffusion.StableDiffusionSafetyChecker",
-        "default": True
+        "default": True,
     },
     {
         "pipeline_action": "controlnet",
         "version": "SD 1.5",
         "category": "stablediffusion",
         "classname": "diffusers.ControlNetModel",
-        "default": True
+        "default": True,
     },
     {
         "pipeline_action": "txt2img",
         "version": "SD 1.5",
         "category": "stablediffusion",
         "classname": "diffusers.AutoPipelineForText2Image",
-        "default": True
+        "default": True,
     },
     {
         "pipeline_action": "txt2img",
         "version": "SD 1.5",
         "category": "controlnet",
         "classname": "diffusers.StableDiffusionControlNetPipeline",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "txt2img",
         "version": "SDXL 1.0",
         "category": "stablediffusion",
         "classname": "diffusers.AutoPipelineForText2Image",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "outpaint",
         "version": "SDXL 1.0",
         "category": "stablediffusion",
         "classname": "diffusers.StableDiffusionXLInpaintPipeline",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "outpaint",
         "version": "SDXL 1.0",
         "category": "controlnet",
         "classname": "diffusers.StableDiffusionXLControlNetInpaintPipeline",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "txt2img",
         "version": "SDXL Turbo",
         "category": "stablediffusion",
         "classname": "diffusers.AutoPipelineForText2Image",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "img2img",
         "version": "SD 1.5",
         "category": "stablediffusion",
         "classname": "diffusers.AutoPipelineForImage2Image",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "img2img",
         "version": "SD 1.5",
         "category": "controlnet",
         "classname": "diffusers.StableDiffusionControlNetImg2ImgPipeline",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "img2img",
         "version": "SDXL 1.0",
         "category": "stablediffusion",
         "classname": "diffusers.AutoPipelineForImage2Image",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "outpaint",
         "version": "SD 1.5",
         "category": "stablediffusion",
         "classname": "diffusers.StableDiffusionInpaintPipeline",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "outpaint",
         "version": "SD 1.5",
         "category": "conrolnet",
         "classname": "diffusers.StableDiffusionControlNetInpaintPipeline",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "inpaint_vae",
         "version": "SD 1.5",
         "category": "stablediffusion",
         "classname": "diffusers.AsymmetricAutoencoderKL",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "text_encoder",
         "version": "SD 1.5",
         "category": "stablediffusion",
         "classname": "transformers.CLIPTextModel",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "text_encoder",
         "version": "SDXL 1.0",
         "category": "stablediffusion",
         "classname": "transformers.CLIPTextModel",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "feature_extractor",
         "version": "SD 1.5",
         "category": "stablediffusion",
         "classname": "transformers.AutoFeatureExtractor",
-        "default": False
+        "default": False,
+    },
+    {
+        "pipeline_action": "txt2img",
+        "version": "Flux S",
+        "category": "stablediffusion",
+        "classname": "diffusers.FluxPipeline",
+        "default": False,
     },
 ]
 
@@ -130,21 +136,21 @@ llm_pipeline_data = [
         "version": "1",
         "category": "llm",
         "classname": "transformers.AutoModelForCausalLM",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "causallm",
         "version": "0.1",
         "category": "llm",
         "classname": "transformers.AutoModelForCausalLM",
-        "default": False
+        "default": False,
     },
     {
         "pipeline_action": "causallm",
         "version": "2",
         "category": "llm",
         "classname": "transformers.AutoModelForCausalLM",
-        "default": False
+        "default": False,
     },
 ]
