+++ src/airunner/enums.py
@@ -334,6 +334,7 @@ class StableDiffusionVersion(Enum):
     SDXL_LIGHTNING = "SDXL Lightning"
     SDXL_HYPER = "SDXL Hyper"
     SD1_5 = "SD 1.5"
+    FLUX_S = "Flux S"
 
 
 class Language(Enum):
@@ -375,6 +376,7 @@ class StatusColors(Enum):
 
 class ModelType(Enum):
     SD = "SD Model"
+    FLUX_MODEL = "Flux Model"
     SD_VAE = "SD VAE"
     SD_UNET = "SD UNet"
     SD_TOKENIZER = "SD Tokenizer"
